mdf/requirements.txt
boxsdk==2.0.0
cryptography==2.7
mdf-connect-client==0.3.1
mdf-forge==0.7.2
mdf-toolbox==0.4.7
PyJWT==1.7.1
requests==2.22.0
requests-toolbelt==0.8.0
urllib3==1.25.3
zipp==0.6.0
pyyaml
pytube
